requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
pandas_datareader==0.10.0
pandas_ta==0.3.14b0
plotly==6.0.0
python_dateutil==2.9.0.post0
scikit_learn==1.3.2
statsmodels==0.14.1
streamlit==1.40.1
ta==0.11.0
yfinance==0.2.52
